feat(nco, axis): fix NCO, AXIS IN, AXIS OUT - update iob_nco module with new clock in domain, fix CDC problems     - NOTE: linux software is outdated - add iob_axistream in and out modules, fix CDC problems     - NOTE: axistream in still needs better CDC for `axis_word_count` to       `NWORDS_rd`. Maybe use gray counter

+++ submodules/LIB/hardware/modules/iob_nco/software/src/iob-nco.h
@@ -6,4 +6,4 @@
 void nco_reset();
 void nco_init(uint32_t base_address);
 void nco_enable(bool enable);
-void nco_set_period(uint32_t period);
+void nco_set_period(uint64_t period);
